+++ POSL/golfers.h
@@ -8,7 +8,7 @@ class Golfers : public Benchmark
 {
     public:
         Golfers(int g, int p, int w);
-        float solutionCost(shared_ptr<Solution> sol);
+        int solutionCost(shared_ptr<Solution> sol);
 
     private:
         int groups, players, weeks;
